(fix): Refactor architecture to support multiple services and improve modularity
