+++ api_tests/core/base_config.py
@@ -1,0 +1,2 @@
+class BaseConfig:
+    BASE_URL: str = ""
