--- api_tests/core/config.py
@@ -1,3 +1,0 @@
-
-class Config:
-    BASE_URL: str = "https://httpbin.org"
