+++ api_tests/services/ozon/config.py
@@ -1,0 +1,5 @@
+from api_tests.core.base_config import BaseConfig
+
+
+class Config(BaseConfig):
+    BASE_URL: str = "https://httpbin.org"
