+++ api_tests/services/ozon/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 
-from api_tests.core.api_client import ApiClient
-from api_tests.core.config import Config
+from api_tests.services.ozon.api_client import ApiClient
+from api_tests.services.ozon.config import Config
 
 pytest_plugins = [
     "api_tests.fixtures.ozon_fixtures",
