+++ docker-compose.yaml
@@ -2,14 +2,13 @@ version: '3.7'
 
 services:
   api-tests-ozon:
-    image: python:3.12-slim
     build:
       context: .
       dockerfile: Dockerfile
+      args:
+        - POETRY_GROUP=ozon
+    image: test-automation:ozon
     volumes:
       - ./:/app
-    command: pytest -s api_tests/tests/ozon_tests
+    command: poetry run pytest -s api_tests/services/ozon/tests
     working_dir: /app
-#    environment:
-#      - CF_ACCESS_CLIENT_ID=${CF_ACCESS_CLIENT_ID}
-#      - CF_ACCESS_CLIENT_SECRET=${CF_ACCESS_CLIENT_SECRET}
