+++ pyproject.toml
@@ -10,12 +10,14 @@ readme = "README.md"
 python = "^3.12"
 pytest = "^8.3.4"
 pydantic = "^2.10.6"
+requests = "^2.32.3"
 
 
 [tool.poetry.group.ozon.dependencies]
-python = "^3.12"
 black = "^22.3"
 pytest = "^8.3.4"
+pydantic = "^2.10.6"
+requests = "^2.32.3"
 
 
 [build-system]
